Fix some compiler warnings

+++ src/mosquitto.c
@@ -189,7 +189,7 @@ static int drop_privileges(struct mosquitto__config *config)
 					}
 				}
 			}
-			if(initgroups(config->user, pwd->pw_gid) == -1){
+			if(initgroups(config->user, (int)pwd->pw_gid) == -1){
 				err = strerror(errno);
 				log__printf(NULL, MOSQ_LOG_ERR, "Error setting groups whilst dropping privileges: %s.", err);
 				return 1;
